ref: extract CsvRecord to its own utility class file, move other types
